+++ src/main/java/com/immobilier/gestionImmobiliere/exceptions/MontantPaiementInvalideException.java
@@ -1,7 +1,9 @@
 package com.immobilier.gestionImmobiliere.exceptions;
 
+import java.math.BigDecimal;
+
 public class MontantPaiementInvalideException extends RuntimeException {
-    public MontantPaiementInvalideException(Double attendu, Double recu) {
+    public MontantPaiementInvalideException(BigDecimal attendu, BigDecimal recu) {
         super("Montant du paiement invalide : attendu " + attendu + ", reçu " + recu);
     }
 }
