+++ src/main/java/com/immobilier/gestionImmobiliere/modules/annonces/dto/requests/UpdateAnnonceDTO.java
@@ -2,6 +2,7 @@ package com.immobilier.gestionImmobiliere.modules.annonces.dto.requests;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Data
@@ -10,6 +11,6 @@ public class UpdateAnnonceDTO {
     private String description;
     private String typeAnnonce;
     private LocalDateTime dateExpiration;
-    private Double prix;
+    private BigDecimal prix;
     private String localisation;
 }
