+++ src/main/java/com/immobilier/gestionImmobiliere/modules/biens/dto/requests/CreateMaisonDTO.java
@@ -3,14 +3,16 @@ package com.immobilier.gestionImmobiliere.modules.biens.dto.requests;
 import jakarta.validation.constraints.NotNull;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class CreateMaisonDTO {
     @NotNull private Integer idCour;
     private String typeMaison;
     private String nomCommunMaison;
     private Integer nombrePiece;
-    private Double loyer;
-    private Double caution;
-    private Double avance;
+    private BigDecimal loyer;
+    private BigDecimal caution;
+    private BigDecimal avance;
     private Integer nombreMoisCaution;
 }
