+++ src/main/java/com/immobilier/gestionImmobiliere/modules/biens/dto/requests/CreerRemboursementDTO.java
@@ -5,9 +5,11 @@ import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Positive;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class CreerRemboursementDTO {
-    @NotNull @Positive private Double montant;
+    @NotNull @Positive private BigDecimal montant;
     @NotBlank private String modeRemboursement;
     private String reference;
     private String motif;
