+++ src/main/java/com/immobilier/gestionImmobiliere/modules/biens/dto/requests/UpdateMaisonDTO.java
@@ -2,12 +2,14 @@ package com.immobilier.gestionImmobiliere.modules.biens.dto.requests;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class UpdateMaisonDTO {
     private String typeMaison;
     private String nomCommunMaison;
     private Integer nombrePiece;
-    private Double loyer;
-    private Double caution;
+    private BigDecimal loyer;
+    private BigDecimal caution;
     private Integer nombreMoisCaution;
 }
