+++ src/main/java/com/immobilier/gestionImmobiliere/modules/biens/dto/responses/MaisonResponseDTO.java
@@ -4,14 +4,16 @@ import com.immobilier.gestionImmobiliere.donnees.biens.model.StatutMaison;
 import lombok.Builder;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data @Builder
 public class MaisonResponseDTO {
     private Integer idMaison;
     private String typeMaison;
     private String nomCommunMaison;
     private Integer nombrePiece;
-    private Double loyer;
-    private Double caution;
+    private BigDecimal loyer;
+    private BigDecimal caution;
     private Integer nombreMoisCaution;
     private StatutMaison statut;
     private Integer idCour;
