+++ src/main/java/com/immobilier/gestionImmobiliere/modules/biens/dto/responses/ModifierDureeLocationResponseDTO.java
@@ -3,10 +3,12 @@ package com.immobilier.gestionImmobiliere.modules.biens.dto.responses;
 import lombok.Builder;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data @Builder
 public class ModifierDureeLocationResponseDTO {
     private LocationBienServiceResponseDTO location;
-    private Double totalEncaisse;
-    private Double totalRembourse;
-    private Double solde;
+    private BigDecimal totalEncaisse;
+    private BigDecimal totalRembourse;
+    private BigDecimal solde;
 }
