+++ src/main/java/com/immobilier/gestionImmobiliere/modules/biens/dto/responses/RemboursementResponseDTO.java
@@ -3,12 +3,13 @@ package com.immobilier.gestionImmobiliere.modules.biens.dto.responses;
 import lombok.Builder;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Data @Builder
 public class RemboursementResponseDTO {
     private Integer idRemboursement;
-    private Double montant;
+    private BigDecimal montant;
     private String modeRemboursement;
     private String reference;
     private String motif;
