+++ src/main/java/com/immobilier/gestionImmobiliere/modules/statistiques/dto/ClientEcheanceDTO.java
@@ -4,6 +4,7 @@ import com.immobilier.gestionImmobiliere.donnees.paiements.model.StatutEcheance;
 import lombok.Builder;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @Data @Builder
@@ -11,7 +12,7 @@ public class ClientEcheanceDTO {
     private Integer idEcheance;
     private String moisLibelle;
     private LocalDate dateEcheance;
-    private Double montantDu;
-    private Double montantPaye;
+    private BigDecimal montantDu;
+    private BigDecimal montantPaye;
     private StatutEcheance statut;
 }
