+++ src/main/java/com/immobilier/gestionImmobiliere/modules/statistiques/dto/ClientRemboursementDTO.java
@@ -3,12 +3,13 @@ package com.immobilier.gestionImmobiliere.modules.statistiques.dto;
 import lombok.Builder;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Data @Builder
 public class ClientRemboursementDTO {
     private Integer idRemboursement;
-    private Double montant;
+    private BigDecimal montant;
     private String motif;
     private LocalDateTime dateRemboursement;
 }
